AddNewtonsoftJson must be called on an environment. Also, some file structure.
